Update README.md
HEALTHCARE PREDICTION ON DIABETIC PATIENTS USING PYTHON is to predict the likelihood of diabetes in patients based on medical data. By using machine learning algorithms, we provide a predictive model that helps healthcare professionals identify high-risk diabetic patients. The goal is to help in early diagnosis and intervention, improving patient care and outcomes.